+++ src/utils.ts
@@ -1,0 +1,3 @@
+export function keys<T extends Object>(obj: T): Array<keyof T> {
+  return Object.keys(obj) as Array<keyof T>;
+}